backend/requirements.txt
# ============================================
# ADMISSIONS COUNSELING CHATBOT - BACKEND DEPENDENCIES
# ============================================

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain & LangGraph
langchain==0.1.4
langchain-openai==0.0.5
langchain-google-genai==0.0.6
langchain-community==0.0.16
langgraph==0.0.20

# Elasticsearch
elasticsearch==8.11.1

# MongoDB
pymongo==4.6.1
motor==3.3.2  # Async MongoDB driver

# Embeddings & ML
sentence-transformers>=2.3.1
torch>=2.0.0
transformers>=4.37.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Environment variables
python-dotenv==1.0.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Data processing
pandas==2.1.4
numpy==1.26.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging
loguru==0.7.2

# CORS
python-cors==1.0.0

# Rate limiting
slowapi==0.1.9

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code quality (optional)
black==23.12.1
flake8==7.0.0
mypy==1.8.0

